+++ Sources/MultiSegmentPicker.swift
@@ -28,7 +28,6 @@ import SwiftUI
     ) {
         _selectedSegmentIndexes = selectedSegmentIndexes
         uiView = MultiSelectSegmentedControl(items: items)
-        uiView.translatesAutoresizingMaskIntoConstraints = false
 
         uiView.allowsMultipleSelection =? allowsMultipleSelection
         uiView.borderWidth =? borderWidth