+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - add SwiftUI view modifiers for appearance properties.
 
+### Fixed
+- make SwiftUI frame correctly respond to `.fixedSize(horizontal: false, vertical: true)`.
+
 ## [2.3.3] - 2020-04-17
 
 ### Fixed